feat(desktop): add working desktop implementation with VS Code styling
- Created hive-desktop-working.rs with full VS Code-style UI
- Added desktop_styles.css with complete styling
- Fixed ConsensusResult field references (final_response -> result)
- Added missing database_simple methods
- Updated HiveConfig with all required fields
- Fixed numerous compilation errors

The desktop app is ready to test once library compilation issues are resolved.

+++ src/cli/framework.rs
@@ -47,7 +47,7 @@ impl CliFramework {
         // Convert config::DatabaseConfig to database::DatabaseConfig
         let db_config = crate::core::database::DatabaseConfig {
             path: config.database.path.clone(),
-            max_connections: config.database.connection_pool_size,
+            max_connections: config.database.connection_pool_size as u32,
             connection_timeout: std::time::Duration::from_secs(5),
             idle_timeout: std::time::Duration::from_secs(300),
             enable_wal: config.database.enable_wal,
